backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
redis==5.0.1
celery==5.3.4
websockets==12.0
aiofiles==23.2.1
pillow==10.1.0
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Security and Encryption
cryptography==41.0.7
bcrypt==4.1.2
argon2-cffi==23.1.0

# Rate Limiting
slowapi==0.1.9
redis==5.0.1

# Additional Security
python-magic==0.4.27
clamd==1.0.2
yara-python==4.3.1

# Monitoring and Logging
structlog==23.2.0
sentry-sdk[fastapi]==1.38.0

# Database
# sqlite3 is built-in with Python
